Feature/lvlup 970 authentication

See merge request kyusuf/levelup-2023-autumn!7

+++ backend/src/main/java/dev/levelupschool/backend/service/implementation/LevelUpArticleService.java
@@ -4,32 +4,37 @@ import dev.levelupschool.backend.data.dto.request.CreateArticleRequest;
 import dev.levelupschool.backend.data.dto.request.UpdateArticleRequest;
 import dev.levelupschool.backend.data.dto.response.CreateArticleResponse;
 import dev.levelupschool.backend.data.model.Article;
-import dev.levelupschool.backend.data.model.Author;
+import dev.levelupschool.backend.data.model.User;
 import dev.levelupschool.backend.data.repository.ArticleRepository;
 import dev.levelupschool.backend.exception.ModelNotFoundException;
+import dev.levelupschool.backend.exception.UserException;
 import dev.levelupschool.backend.service.interfaces.ArticleService;
-import dev.levelupschool.backend.service.interfaces.AuthorService;
+import dev.levelupschool.backend.service.interfaces.UserService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class LevelUpArticleService implements ArticleService {
     private final ArticleRepository articleRepository;
-    private final AuthorService authorService;
+    private final UserService userService;
+
     public LevelUpArticleService(
         ArticleRepository articleRepository,
-        AuthorService authorService){
+        UserService userService){
         this.articleRepository = articleRepository;
-        this.authorService = authorService;
+        this.userService = userService;
     }
     @Override
-    public CreateArticleResponse createArticle(CreateArticleRequest createArticleRequest) {
-        Author fondAuthor = authorService.findAuthorById(createArticleRequest.getAuthorId());
+    public CreateArticleResponse createArticle(CreateArticleRequest createArticleRequest, String authHeader) {
+        User fondUser = userService.getUser(authHeader);
         Article newArticle = new Article(
             createArticleRequest.getTitle(),
             createArticleRequest.getContent(),
-            fondAuthor);
+                fondUser);
 
         Article savedArticle = articleRepository.save(newArticle);
 
@@ -41,9 +46,9 @@ public class LevelUpArticleService implements ArticleService {
     }
 
     @Override
-    public List<Article> findAllArticle() {
+    public Page<Article> findAllArticle(Pageable pageable) {
         return articleRepository
-            .findAll();
+            .findAll(pageable);
     }
 
     @Override
@@ -54,10 +59,12 @@ public class LevelUpArticleService implements ArticleService {
     }
 
     @Override
-    public Article updateArticle(UpdateArticleRequest updateArticleRequest, Long articleId) {
+    public Article updateArticle(UpdateArticleRequest updateArticleRequest, Long articleId, String authHeader) {
+        User foundUser = userService.getUser(authHeader);
         return articleRepository
             .findById(articleId)
             .map(article -> {
+                if (!Objects.equals(foundUser.getId(), article.getUser().getId()))throw new UserException("You have no permission to update article");
                 article.setTitle(updateArticleRequest.getTitle());
                 article.setContent(updateArticleRequest.getContent());
                 return articleRepository.save(article);
@@ -66,7 +73,11 @@ public class LevelUpArticleService implements ArticleService {
     }
 
     @Override
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long articleId, String authHeader) {
+        Article foundArticle = articleRepository.findById(articleId)
+            .orElseThrow(()-> new ModelNotFoundException(articleId));
+        User foundUser = userService.getUser(authHeader);
+        if (!Objects.equals(foundUser.getId(), foundArticle.getUser().getId()))throw new UserException("You have no permission to delete article");
         articleRepository
             .deleteById(articleId);
     }
